Alterando a view e aplicando regras para as requisições

+++ transacoes/models.py
@@ -5,7 +5,8 @@ from django.db import models
 
 class Receitas(models.Model):
     descricao = models.CharField(max_length=200)
-    valor = models.FloatField(null=False, blank=False)
+    valor = models.DecimalField(
+        decimal_places=2, max_digits=8, null=False, blank=False)
     data = models.DateField(null=False, blank=False)
 
     def __str__(self):
@@ -14,7 +15,8 @@ class Receitas(models.Model):
 
 class Despesas(models.Model):
     descricao = models.CharField(max_length=200)
-    valor = models.FloatField(null=False, blank=False)
+    valor = models.DecimalField(
+        decimal_places=2, max_digits=15, null=False, blank=False)
     data = models.DateField(null=False, blank=False)
 
     def __str__(self):
